+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Head (0.7.3)
 * coding guideline update: csscomb, convert tabs to spaces (based on codeguide.co)
 * add the most important attributes to hamburger at nav.scss
+* added marker.click preset to googlemaps template
 * renaming versioning to WPSeed
 * cleanup